requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
beautifulsoup4>=4.12.2
langchain>=0.0.300
openai>=0.27.0
sentence-transformers>=2.2.2
transformers>=4.35.2,<5.0.0
huggingface-hub>=0.33.4,<1.0.0
faiss-cpu>=1.7.4   # optional; use faiss-cpu or faiss-cpu-binary per platform
python-dotenv>=1.0.0
typing-extensions>=4.8.0
langchain_community
langchain_huggingface
langchain_openai
paramiko